+++ images/obsidian_sword.png
